refactor: move FactoryLinkProvider to providers directory
