+++ src/test/suite/extension.test.ts
@@ -1,6 +1,6 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
-import { FactoryLinkProvider } from "../../extension";
+import { FactoryLinkProvider } from "../../providers/factoryLinkProvider";
 import * as path from "path";
 import * as os from "os";
 import * as fs from "fs";
